requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.10.0
apimatic-core==0.2.21
apimatic-core-interfaces==0.1.6
apimatic-requests-client-adapter==0.1.7
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
CacheControl==0.12.14
cairocffi==1.7.1
CairoSVG==2.8.0
celery==5.4.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cssselect2==0.8.0
defusedxml==0.7.1
deprecation==2.1.0
Django==5.2.1
django-celery-beat==2.8.0
django-celery-results==2.0.1
django-dotenv==1.4.2
django-jazzmin==3.0.1
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
django-timezone-field==7.1
django-widget-tweaks==1.5.0
et_xmlfile==2.0.0
exceptiongroup==1.3.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jsonpickle==3.3.0
jsonpointer==2.4
kombu==5.5.4
msgpack==1.1.1
openpyxl==3.0.7
packaging==25.0
Pillow==9.5.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pyphen==0.17.2
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
square==0.0.3
squareup==36.0.0.20240417
stripe==12.0.1
tinycss2==1.4.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
WeasyPrint==50
webencodings==0.5.1
